requirements.txt
aiohttp
aiohttp_cors
aioredis
python-socketio
pyjwt
bcrypt